tests/requirements.txt
pytest==7.4.3
pyspark==3.5.0
delta-spark==3.0.0
great-expectations==0.18.8
pyyaml==6.0.1
requests==2.31.0
